Add interfaces and models for image assets and related functionality
- Created IEntity, IImageFile, ISoundFile, ITimeStamped, IUsageStats, and IVideoFile interfaces.
- Implemented ImageAsset model with properties for image details, timestamps, and usage statistics.
- Added Razor Pages for managing image assets, including an index page.
- Included a new XML key file for data protection.

+++ Models/SoundMissing.cs
@@ -2,11 +2,11 @@ namespace TinyGenerator.Models;
 
 public sealed class SoundMissing
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Type { get; set; } = "fx"; // fx, amb, music
     public string Prompt { get; set; } = string.Empty;
     public string? Tags { get; set; }
-    public long? StoryId { get; set; }
+    public int? StoryId { get; set; }
     public string? StoryTitle { get; set; }
     public string? Source { get; set; }
     public int Occurrences { get; set; }
